Backend/requirements.txt

fastapi
uvicorn
motor
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
email-validator
pytest
requests
